Separates out encoding from the structure of a Hamiltonian.
This is necessary for later work to allow encodings to be used with arbitrary Hamiltonians.

Firstly, a new submodule, Hamiltonians is created and functionality from the base FermionQubitEncoding has been moved into there.
These functions have been removed from the base encoding, which means that hamiltonain coefficients are no longer needed at init. Instead a base encoding needs a number of qubits and a number of modes.

For convenience a `from_hamiltonian_coefficients` class function has been added to TernaryTree.

Other than that, a lot of work had to be done to make the tests happy. Some of these have been moved to `test_hamiltonians`.