Separate Hamiltonians from Encodings (#24)
# Arbitrary Encodings
Initially, encodings were built with the assumption of encoding the electronic structure hamiltonian.

These have now been separated out, so that encodings can be applied to arbitrary fermionic operators. At the moment, only the ESH is provided, though it is now possible to add more general functions fairly easily.

There are two distinct ways hamiltonians can be encoded: Templates and enumerated encodings. 

Templates do not require hamiltonian coefficients, requiring only an encoding and the signature of fermionic operators incuded in the Hamiltonian. They're returned as dict of byte-hashed symplectic vectors and coefficients, so that they can be easily indexed, transformed and, later, opetimized over. 

In contrast, when a mode_op_map and hamiltonain coefficients are provided, a full Pauli-string Hamiltonian can be generated.

Some information on using ferrmions encoding methods with other qSDKs has been added to the docs. More of these, and full examples, will be added in another PR.