Huffman Ternary Tree (#48)
Huffman-code ternary trees are designed to reduce the coefficient Pauli-weight of a Hamiltonian for use in QDRIFT. 

The [original paper](https://iopscience.iop.org/article/10.1088/0256-307X/42/10/100001) describes a method to generate non-vaccum-preserving TTs, but only vaccum-preserving has been implemented.

This results in a generally higher coefficent Pauli-weight than described in the paper.