BUG: symplectic_to_sparse (#55)
symplectic_to_sparse was not carrying through ipowers correctly, leading to correct calculations of symplectic terms, but wrong output.

This has been corrected, along with standardising the conversion functions

    symplectic_to_pauli
    pauli_to_symplectic
    symplectic_to_sparse
    so that they all take in an ipower, and return an updated ipower.

An alternative could have been to alter the definition of symplectic form so that x=True, z=True is equivalent to Y, rather than XZ=iY. We'll still with WYSIWYG, as another form would require significant warnings in documentation and commenting.