Hamiltonian Adaptive Ternary Tree (#66)
Add Hamiltonian Adaptvie Ternary Tree.

To make this easy, added granulaity is needed in enumeration. Rather than enumerate nodes and apply the stirng paring algorithm, it needs to be possible to enumerate branches as majorana operators. This can be done by applying the string pairing algorithm or explcitly for HATT/Huffman.