allow seting n_qubits in TernaryTree and make bonsai deterministic (#87)
When creating a TernaryTree from a device connectivity graph, there was an issue if the qubit labels used were not a contiguous list beginning at zero, in that when creating the symplectic matrix, we would be trying to set an operator for a qubit index which was out of bounds.

To get around this, we allow setting the number of qubits in the encoding manually, so that a numbe equal to the total on a device can be used. Any qubits which are not involved in the encoding are padded with zeros.

Additionally, the bonsai algorithm was non-deterministic as rustworkx was returning neighbor nodes in arbitrary order. This was leading to non-deterministic test results! So neighbours are now sorted by node index before use.