Merge branch 'main' into rust-topphatt (#95)
Adds the ability to create encodings with a number of qubits larger than the number of fermionic modes, additionally ensures that input qubit indices are used when building encodings.

This is necessary for the bonsai algorithm, where we want to use a subset of qubits ona. device and require the encoding to use specific qubits.

The way to do this is to pass in a TTFlatpack which uses the qubit indices required, then run build_encoding with at least as many qubits as the maximum number in the flatpack.